+++ internal/provider/data_source_iosxe_interface_vlan_test.go
@@ -25,6 +25,10 @@ func TestAccDataSourceIosxeInterfaceVLAN(t *testing.T) {
 					resource.TestCheckResourceAttr("data.iosxe_interface_vlan.test", "ipv4_address", "10.1.1.1"),
 					resource.TestCheckResourceAttr("data.iosxe_interface_vlan.test", "ipv4_address_mask", "255.255.255.0"),
 					resource.TestCheckResourceAttr("data.iosxe_interface_vlan.test", "ip_dhcp_relay_source_interface", "Loopback100"),
+					resource.TestCheckResourceAttr("data.iosxe_interface_vlan.test", "ip_access_group_in", "1"),
+					resource.TestCheckResourceAttr("data.iosxe_interface_vlan.test", "ip_access_group_in_enable", "true"),
+					resource.TestCheckResourceAttr("data.iosxe_interface_vlan.test", "ip_access_group_out", "1"),
+					resource.TestCheckResourceAttr("data.iosxe_interface_vlan.test", "ip_access_group_out_enable", "true"),
 					resource.TestCheckResourceAttr("data.iosxe_interface_vlan.test", "helper_addresses.0.address", "10.10.10.10"),
 					resource.TestCheckResourceAttr("data.iosxe_interface_vlan.test", "helper_addresses.0.global", "false"),
 					resource.TestCheckResourceAttr("data.iosxe_interface_vlan.test", "helper_addresses.0.vrf", "VRF1"),
@@ -65,6 +69,10 @@ resource "iosxe_interface_vlan" "test" {
   ipv4_address = "10.1.1.1"
   ipv4_address_mask = "255.255.255.0"
   ip_dhcp_relay_source_interface = "Loopback100"
+  ip_access_group_in = "1"
+  ip_access_group_in_enable = true
+  ip_access_group_out = "1"
+  ip_access_group_out_enable = true
   helper_addresses = [{
     address = "10.10.10.10"
     global = false