+++ internal/provider/model_iosxe_interface_vlan.go
@@ -29,6 +29,10 @@ type InterfaceVLAN struct {
 	Ipv4AddressMask            types.String                   `tfsdk:"ipv4_address_mask"`
 	Unnumbered                 types.String                   `tfsdk:"unnumbered"`
 	IpDhcpRelaySourceInterface types.String                   `tfsdk:"ip_dhcp_relay_source_interface"`
+	IpAccessGroupIn            types.String                   `tfsdk:"ip_access_group_in"`
+	IpAccessGroupInEnable      types.Bool                     `tfsdk:"ip_access_group_in_enable"`
+	IpAccessGroupOut           types.String                   `tfsdk:"ip_access_group_out"`
+	IpAccessGroupOutEnable     types.Bool                     `tfsdk:"ip_access_group_out_enable"`
 	HelperAddresses            []InterfaceVLANHelperAddresses `tfsdk:"helper_addresses"`
 }
 type InterfaceVLANHelperAddresses struct {
@@ -82,6 +86,22 @@ func (data InterfaceVLAN) toBody(ctx context.Context) string {
 	}
 	if !data.IpDhcpRelaySourceInterface.Null && !data.IpDhcpRelaySourceInterface.Unknown {
 		body, _ = sjson.Set(body, helpers.LastElement(data.getPath())+"."+"ip.dhcp.Cisco-IOS-XE-dhcp:relay.source-interface", data.IpDhcpRelaySourceInterface.Value)
+	}
+	if !data.IpAccessGroupIn.Null && !data.IpAccessGroupIn.Unknown {
+		body, _ = sjson.Set(body, helpers.LastElement(data.getPath())+"."+"ip.access-group.in.acl.acl-name", data.IpAccessGroupIn.Value)
+	}
+	if !data.IpAccessGroupInEnable.Null && !data.IpAccessGroupInEnable.Unknown {
+		if data.IpAccessGroupInEnable.Value {
+			body, _ = sjson.Set(body, helpers.LastElement(data.getPath())+"."+"ip.access-group.in.acl.in", map[string]string{})
+		}
+	}
+	if !data.IpAccessGroupOut.Null && !data.IpAccessGroupOut.Unknown {
+		body, _ = sjson.Set(body, helpers.LastElement(data.getPath())+"."+"ip.access-group.out.acl.acl-name", data.IpAccessGroupOut.Value)
+	}
+	if !data.IpAccessGroupOutEnable.Null && !data.IpAccessGroupOutEnable.Unknown {
+		if data.IpAccessGroupOutEnable.Value {
+			body, _ = sjson.Set(body, helpers.LastElement(data.getPath())+"."+"ip.access-group.out.acl.out", map[string]string{})
+		}
 	}
 	if len(data.HelperAddresses) > 0 {
 		body, _ = sjson.Set(body, helpers.LastElement(data.getPath())+"."+"ip.helper-address", []interface{}{})
@@ -151,6 +171,26 @@ func (data *InterfaceVLAN) updateFromBody(ctx context.Context, res gjson.Result)
 		data.IpDhcpRelaySourceInterface.Value = value.String()
 	} else {
 		data.IpDhcpRelaySourceInterface.Null = true
+	}
+	if value := res.Get(prefix + "ip.access-group.in.acl.acl-name"); value.Exists() {
+		data.IpAccessGroupIn.Value = value.String()
+	} else {
+		data.IpAccessGroupIn.Null = true
+	}
+	if value := res.Get(prefix + "ip.access-group.in.acl.in"); value.Exists() {
+		data.IpAccessGroupInEnable.Value = true
+	} else {
+		data.IpAccessGroupInEnable.Value = false
+	}
+	if value := res.Get(prefix + "ip.access-group.out.acl.acl-name"); value.Exists() {
+		data.IpAccessGroupOut.Value = value.String()
+	} else {
+		data.IpAccessGroupOut.Null = true
+	}
+	if value := res.Get(prefix + "ip.access-group.out.acl.out"); value.Exists() {
+		data.IpAccessGroupOutEnable.Value = true
+	} else {
+		data.IpAccessGroupOutEnable.Value = false
 	}
 	for i := range data.HelperAddresses {
 		keys := [...]string{"address"}
@@ -236,6 +276,28 @@ func (data *InterfaceVLAN) fromBody(ctx context.Context, res gjson.Result) {
 		data.IpDhcpRelaySourceInterface.Value = value.String()
 		data.IpDhcpRelaySourceInterface.Null = false
 	}
+	if value := res.Get(prefix + "ip.access-group.in.acl.acl-name"); value.Exists() {
+		data.IpAccessGroupIn.Value = value.String()
+		data.IpAccessGroupIn.Null = false
+	}
+	if value := res.Get(prefix + "ip.access-group.in.acl.in"); value.Exists() {
+		data.IpAccessGroupInEnable.Value = true
+		data.IpAccessGroupInEnable.Null = false
+	} else {
+		data.IpAccessGroupInEnable.Value = false
+		data.IpAccessGroupInEnable.Null = false
+	}
+	if value := res.Get(prefix + "ip.access-group.out.acl.acl-name"); value.Exists() {
+		data.IpAccessGroupOut.Value = value.String()
+		data.IpAccessGroupOut.Null = false
+	}
+	if value := res.Get(prefix + "ip.access-group.out.acl.out"); value.Exists() {
+		data.IpAccessGroupOutEnable.Value = true
+		data.IpAccessGroupOutEnable.Null = false
+	} else {
+		data.IpAccessGroupOutEnable.Value = false
+		data.IpAccessGroupOutEnable.Null = false
+	}
 	if value := res.Get(prefix + "ip.helper-address"); value.Exists() {
 		data.HelperAddresses = make([]InterfaceVLANHelperAddresses, 0)
 		value.ForEach(func(k, v gjson.Result) bool {
@@ -303,6 +365,22 @@ func (data *InterfaceVLAN) setUnknownValues(ctx context.Context) {
 		data.IpDhcpRelaySourceInterface.Unknown = false
 		data.IpDhcpRelaySourceInterface.Null = true
 	}
+	if data.IpAccessGroupIn.Unknown {
+		data.IpAccessGroupIn.Unknown = false
+		data.IpAccessGroupIn.Null = true
+	}
+	if data.IpAccessGroupInEnable.Unknown {
+		data.IpAccessGroupInEnable.Unknown = false
+		data.IpAccessGroupInEnable.Null = true
+	}
+	if data.IpAccessGroupOut.Unknown {
+		data.IpAccessGroupOut.Unknown = false
+		data.IpAccessGroupOut.Null = true
+	}
+	if data.IpAccessGroupOutEnable.Unknown {
+		data.IpAccessGroupOutEnable.Unknown = false
+		data.IpAccessGroupOutEnable.Null = true
+	}
 	for i := range data.HelperAddresses {
 		if data.HelperAddresses[i].Address.Unknown {
 			data.HelperAddresses[i].Address.Unknown = false
@@ -353,6 +431,12 @@ func (data *InterfaceVLAN) getEmptyLeafsDelete(ctx context.Context) []string {
 	emptyLeafsDelete := make([]string, 0)
 	if !data.Shutdown.Value {
 		emptyLeafsDelete = append(emptyLeafsDelete, fmt.Sprintf("%v/shutdown", data.getPath()))
+	}
+	if !data.IpAccessGroupInEnable.Value {
+		emptyLeafsDelete = append(emptyLeafsDelete, fmt.Sprintf("%v/ip/access-group/in/apply-type/apply-intf/acl/in", data.getPath()))
+	}
+	if !data.IpAccessGroupOutEnable.Value {
+		emptyLeafsDelete = append(emptyLeafsDelete, fmt.Sprintf("%v/ip/access-group/out/apply-type/apply-intf/acl/out", data.getPath()))
 	}
 
 	for i := range data.HelperAddresses {